backend/pyproject.toml
[tool.poetry]
name = "critter"
version = "0.1.0"
description = "Twitter for critters"
authors = ["stromjust <justas.stromilas@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.23"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
psycopg2 = "^2.9.1"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
